Refactoring web ui
