+++ replica-devtools-client/src/main/kotlin/me/aartikov/replica/devtools/client/Body.kt
@@ -1,0 +1,79 @@
+package me.aartikov.replica.devtools.client
+
+import androidx.compose.runtime.Composable
+import me.aartikov.replica.devtools.dto.ReplicaClientDto
+import org.jetbrains.compose.web.css.*
+import org.jetbrains.compose.web.css.keywords.auto
+import org.jetbrains.compose.web.dom.Div
+import org.jetbrains.compose.web.dom.Ul
+
+@Composable
+fun Body(state: ReplicaClientDto) {
+    Card(
+        attrs = {
+            style {
+                position(Position.Absolute)
+                width(90.percent)
+                height(100.percent)
+                top(0.px)
+                bottom(0.px)
+                left(0.px)
+                right(0.px)
+                property("margin", auto)
+            }
+        }
+    ) {
+        Div(
+            attrs = {
+                style {
+                    width(100.percent)
+                    height(100.percent)
+                    position(Position.Absolute)
+                    top(0.px)
+                    left(0.px)
+                }
+            }
+        ) {
+            Content(state)
+        }
+    }
+}
+
+@Composable
+fun Content(state: ReplicaClientDto) {
+    Div(
+        attrs = {
+            style {
+                width(100.percent)
+                height(100.percent)
+                display(DisplayStyle.Flex)
+                flexFlow(FlexDirection.Column, FlexWrap.Nowrap)
+                overflowY("scroll")
+            }
+        }
+    ) {
+        Div(
+            attrs = {
+                style {
+                    width(100.percent)
+                    property("flex", "0 1 auto")
+                }
+            }
+        ) { NavBar(title = "Replica dev tool") }
+        Ul(
+            attrs = {
+                style {
+                    width(100.percent)
+                    margin(0.px)
+                }
+            }
+        ) {
+            state.replicas.values.forEach { replica ->
+                ReplicaItemUi(item = replica)
+            }
+            state.keyedReplicas.values.forEach { replica ->
+                KeyedReplicaItem(item = replica)
+            }
+        }
+    }
+}
